requirements.txt
# SGB V RAG System - Python Dependencies

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==2.1.0

# Data processing
numpy==1.24.3
pandas==2.1.3

# Vector database and search
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Embeddings and transformers
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.35.2

# LLM Integration
openai==1.3.8
anthropic==0.7.1
langchain==0.1.1
langchain-openai==0.0.5
langchain-community==0.0.8

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web UI
streamlit==1.28.1
streamlit-option-menu==0.3.6

# Environment and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
python-logging-loki==0.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development tools
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
